tools/syslog-server/requirements.txt
# Syslog Server - Requirements

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.6.0
pydantic-settings==2.2.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Cache
redis==5.0.1
hiredis==2.3.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Monitoring
prometheus-client==0.19.0

# Logging
structlog==24.1.0

# Geo Location
geoip2==4.8.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
click==8.1.7
rich==13.7.0
